requirements.txt
pydantic>=1.8.0,<2.0.0
PyJWT==1.7.1
uvicorn>0.21.0
fastapi==0.61.0
fastapi-jwt-auth==0.5.0
fastapi-users==3.1.2
bcrypt==4.0.1
passlib==1.7.4